+++ test/data/sample_tests/require_annotations/annotations.js
@@ -1,0 +1,5 @@
+/**
+ * @venus-library mocha
+ */
+
+var foo = 1;